Summarise a dataframe

A tidyverse-friendly summary function that summarises a dataframe by column type.

my_skim(data, ...)

Arguments

data
A tibble, or an object that can be coerced into a tibble.

...
Columns to select for skimming. When none are provided, the default is to skim all columns.

Details

Works with dplyr::group_by() and the pipe. See the skim documentation for more details. Adapts the skimr::skim() function to include proportion counts for factor variables

In general, more informative results are returned if character-type columns are first converted to factors (see examples below)

Examples

# summarise the iris dataset
my_skim(iris)

#> ── Data Summary ────────────────────────
#>                            Values
#> Name                       iris  
#> Number of rows             150   
#> Number of columns          5     
#> _______________________          
#> Column type frequency:           
#>   factor                   1     
#>   numeric                  4     
#> ________________________         
#> Group variables            None  
#> 
#> ── Variable type: factor ───────────────────────────────────────────────────────
#>   skim_variable n_missing complete_rate ordered n_unique
#> 1 Species               0             1 FALSE          3
#>   top_counts                pct                                                
#> 1 set: 50, ver: 50, vir: 50 setosa: 33.3%, versicolor: 33.3%, virginica: 33.3%,
#> 
#> ── Variable type: numeric ──────────────────────────────────────────────────────
#>   skim_variable n_missing complete_rate  mean    sd    p0   p25   p50   p75
#> 1 Sepal.Length          0             1  5.84 0.828   4.3   5.1  5.8    6.4
#> 2 Sepal.Width           0             1  3.06 0.436   2     2.8  3      3.3
#> 3 Petal.Length          0             1  3.76 1.77    1     1.6  4.35   5.1
#> 4 Petal.Width           0             1  1.20 0.762   0.1   0.3  1.3    1.8
#>    p100 hist 
#> 1   7.9 ▆▇▇▅▂
#> 2   4.4 ▁▆▇▂▁
#> 3   6.9 ▇▁▆▇▂
#> 4   2.5 ▇▁▇▅▃

# summarise the mtcars dataset by transmissions type ("am": 0 = automatic, 1 = manual)
library(magrittr)
mtcars %>%
  dplyr::mutate(
    dplyr::across(
      tidyselect::all_of(c("cyl", "vs", "am", "gear", "carb")),
      as.factor)
    ) %>%
  dplyr::group_by(am) %>%
  my_skim()

#> ── Data Summary ────────────────────────
#>                            Values    
#> Name                       Piped data
#> Number of rows             32        
#> Number of columns          11        
#> _______________________              
#> Column type frequency:               
#>   factor                   4         
#>   numeric                  6         
#> ________________________             
#> Group variables            am        
#> 
#> ── Variable type: factor ───────────────────────────────────────────────────────
#>   skim_variable am    n_missing complete_rate ordered n_unique
#> 1 cyl           0             0             1 FALSE          3
#> 2 cyl           1             0             1 FALSE          3
#> 3 vs            0             0             1 FALSE          2
#> 4 vs            1             0             1 FALSE          2
#> 5 gear          0             0             1 FALSE          2
#> 6 gear          1             0             1 FALSE          2
#> 7 carb          0             0             1 FALSE          4
#> 8 carb          1             0             1 FALSE          5
#>   top_counts             pct                                                   
#> 1 8: 12, 6: 4, 4: 3      4: 15.8%, 6: 21.1%, 8: 63.2%,                         
#> 2 4: 8, 6: 3, 8: 2       4: 61.5%, 6: 23.1%, 8: 15.4%,                         
#> 3 0: 12, 1: 7            0: 63.2%, 1: 36.8%,                                   
#> 4 1: 7, 0: 6             0: 46.2%, 1: 53.8%,                                   
#> 5 3: 15, 4: 4, 5: 0      3: 78.9%, 4: 21.1%, 5: 0%,                            
#> 6 4: 8, 5: 5, 3: 0       3: 0%, 4: 61.5%, 5: 38.5%,                            
#> 7 4: 7, 2: 6, 1: 3, 3: 3 1: 15.8%, 2: 31.6%, 3: 15.8%, 4: 36.8%, 6: 0%, 8: 0%, 
#> 8 1: 4, 2: 4, 4: 3, 6: 1 1: 30.8%, 2: 30.8%, 3: 0%, 4: 23.1%, 6: 7.7%, 8: 7.7%,
#> 
#> ── Variable type: numeric ──────────────────────────────────────────────────────
#>    skim_variable am    n_missing complete_rate   mean      sd     p0    p25
#>  1 mpg           0             0             1  17.1    3.83   10.4   15.0 
#>  2 mpg           1             0             1  24.4    6.17   15     21   
#>  3 disp          0             0             1 290.   110.    120.   196.  
#>  4 disp          1             0             1 144.    87.2    71.1   79   
#>  5 hp            0             0             1 160.    53.9    62    116.  
#>  6 hp            1             0             1 127.    84.1    52     66   
#>  7 drat          0             0             1   3.29   0.392   2.76   3.07
#>  8 drat          1             0             1   4.05   0.364   3.54   3.85
#>  9 wt            0             0             1   3.77   0.777   2.46   3.44
#> 10 wt            1             0             1   2.41   0.617   1.51   1.94
#> 11 qsec          0             0             1  18.2    1.75   15.4   17.2 
#> 12 qsec          1             0             1  17.4    1.79   14.5   16.5 
#>       p50    p75   p100 hist 
#>  1  17.3   19.2   24.4  ▂▇▇▅▂
#>  2  22.8   30.4   33.9  ▃▇▃▂▇
#>  3 276.   360    472    ▇▃▇▆▅
#>  4 120.   160    351    ▇▃▁▁▂
#>  5 175    192.   245    ▃▅▂▇▅
#>  6 109    113    335    ▇▅▁▁▁
#>  7   3.15   3.70   3.92 ▂▇▁▁▅
#>  8   4.08   4.22   4.93 ▅▇▅▂▂
#>  9   3.52   3.84   5.42 ▁▇▃▁▂
#> 10   2.32   2.78   3.57 ▆▇▂▆▃
#> 11  17.8   19.2   22.9  ▂▇▂▂▁
#> 12  17.0   18.6   19.9  ▇▂▇▇▇